add regression test for JSX element wrapping at the 80-col boundary
Summary:
`list` generates two `Sequence` nodes, one inside the other. if the `wrap` is a `flat_space`, and the items are exactly the same width as the remaining space on the line, then the outer sequence thinks that a break is necessary so it doesn't print the `flat_space`, but then the inner sequence thinks it fits and doesn't break -- leading to a missing space.

in this case, it manifests as `<foo bar ...>` being printed as `<foobar ...>` instead of `<foo\n  bar ...>`

this diff adds a test demonstrating the current broken behavior, so that a followup can show the fix.

Reviewed By: pvdz

Differential Revision: D10258579

fbshipit-source-id: 56ca0d186ca76d0ddde90629760b7b2894df767e