Use Nel.t for Missing Annotation traces
Summary:
We were getting confusing crashes in flow. gabelevi found that the offender was the call to List.tl and List.hd in the assert_ground visitor trace collection logic.

The root cause was that the boolean "should_add" was not being used correctly-- it was adding when it should not and not adding when it should. Later on, we would pop a reason from the reason stack if we should have added one. Since the logic was reversed, this wasn't happening.

Since this is a debug tool, we can remove the logic that makes this feature convoluted in the first place, which only tries to minimize the error stack. In addition to being more complicated than necessary, it's also easy to produce a confusing trace using this scheme, since we were removing the middle reasons whenever we got past the limit. Instead, we now will only add reasons to the trace until we hit the max limit, as we do with other traces.

As well, we can augment `Nel` a bit to enforce the invariant we need via the type system. We also call out our invariants explicitly that we rely on when using functions that throw here.

Reviewed By: avikchaudhuri

Differential Revision: D10263761

fbshipit-source-id: 23ef6dc125becc6aa23fa6b8ed068470d676ede6