ALoc.compare
Summary:
Even abstract locations need to have a well defined order, as ordering can be significant to type checking (for better or worse). A non-stable order would lead to unpredictable behavior, e.g., errors added on recheck, existing errors lost on recheck. For now, `Loc.compare` on concrete locations serves this purpose.

In the future, abstract locations will also define a stable order, based first on the `source`, which we will still store concretely, and then on some property of the keys we store for locations. It is highly likely that this order will be different than the order of the corresponding concrete locations, but it will be stable.

Reviewed By: samwgoldman

Differential Revision: D10306367

fbshipit-source-id: 3e8aaa95dbd1d45d2cb795ba26e476e7246cbf17