Parameterize error types over location
Summary:
This was already done for some of the types in this module, but it was far from complete. This finishes the job.

The plan is to store messages containing `ALoc.t`s, and then convert them to messages containing `Loc.t`s for printing. It's quite a bit cleaner than converting locations ad-hoc during the printing process.

Reviewed By: panagosg7

Differential Revision: D10306368

fbshipit-source-id: c293cfed06b34c1079f38072268d4b4b4d43e302