[parser] add inexact flag to indicate presence of ... on Ast.Type.Object.t
Summary:
We set `inexact = true` when we encounter the explicit inexact syntax introduced in D10220687.

The current way to determine if an object is exact is by looking at the `exact` field on the Ast node. However, if `exact = false` there is no way to determine if the object actually explicitly indicated that it is an inexact object with `...`. We need this functionality in order to implement a lint to complain when there is no `...` in an inexact object.

We should be able to remove this after all objects are exact by default and inexact with `...`.

In a follow-up diff, I'll be adding `inexact` to the external ast so it may be consumed by jscodeshift in codemods to switch the syntax.

Reviewed By: samwgoldman

Differential Revision: D10361110

fbshipit-source-id: 0ab5c464d67c385fe7092f304f2696440b037f61