[parser] add inexact flag on objects to estree printed ast
Summary:
Adds the inexact flag on `ObjectTypeAnnotation`s to the estree AST. This will allow us to determine if "..." is present in inexact objects in tools that use the flow AST.

When inexact is present, we can be sure the `ObjectTypeAnnotation` is actually representing an object type. If it is true, then the object type uses `...`, otherwise it does not.

Support for this should be added to babel, prettier, and recast.

Reviewed By: samwgoldman

Differential Revision: D10361894

fbshipit-source-id: 5a274a5433eba5c1c920b15e60b1250b2c06acb0