[hack] fix control flow in naming phase
Summary:
This is a first of a series of diffs. It is not meant to be shipped as is since it does not deal with a few corner cases (unsafe blocks, isset($x)...), and creates a lot of dead code in naming.ml, but I put it up for review already.

In this diff, we silence the "undefined" error issued in naming and instead generate this error in typing.ml, whenever looking up the Next continuation fails.

Since the way naming generates integer ids for locals is broken, we now stop generating unique integer ids in most cases - except for now for `$$` (used with the pipe operator). Meaning all variables have id 0 and Local_id.compare now also takes into account the string id of locals.

As a result of this diff, a lot of code in naming will become useless and will be cleaned up in a later diff.

Reviewed By: manzyuk

Differential Revision: D7763002

fbshipit-source-id: 8e2dba85fe1cc9c822b1ab3e804233199df7d360