[tool] skip curses interface in `add-comments --all`
Summary: using `--all` previously fired up the curses interface and then simulated selecting all the errors. `--all` really only makes sense if you're calling it from a script, and you don't really want the curses interface in a script... it breaks the console output from the rest of the script.

Reviewed By: jbrown215

Differential Revision: D10380572

fbshipit-source-id: 289147b1e8ee0fecf58284f48ccf33b9152a1e71