change exports maps in file sig to lists to accommodate overloading
Summary:
When there are multiple exports of the same name, a map keeps only the last one. This breaks
overloading at the top level.

Maps also helped relate an `export` with its corresponding `export_info`. To preserve this relation,
we ensire that the corresponding lists are parallel. This also results in some other
simplifications, and more cleanup is possible by tightening data structures further.

Reviewed By: panagosg7

Differential Revision: D10280157

fbshipit-source-id: e17e77e8a7c13761f57b0e75eeda6ab1570cd2b8