array summarization
Summary:
Previously, the element type of an array was considered to be the union of the types of its
elements. Instead, we can try to compute a less precise, more "natural" join. (Precision can
actually hurt in invariant locations.)

We lift number literal types to the number type, and so forth.

The join of a pair of array types whose element types have already been joined is an array type
whose element type is the join of the pair of element types.

The join of a pair of object types is an object type whose property types are the joins of the
corresponding pairs of property types; the keysets of the pair of object types must be the same.

Any other join falls back to the union.

A couple of other ideas were considered, but didn't make it into this diff: joining arrays of the
same length as tuples; and joining objects that contain uniform property types as dictionaries.

Reviewed By: panagosg7

Differential Revision: D10327705

fbshipit-source-id: abecc657232d3c530324b9eb4e74b06ccf2c3aef