Make Error mapper polymorphic
Summary:
As requested in D10306366, I am reusing the existing mapper for conversions from `ALoc.t` to `Loc.t`, rather than defining a separate one that is polymorphic.

The upside to this is that there is less code to maintain. The downside is that because we are returning a different type, the type system will not allow us to check for equality and return the original node, therefore avoiding extra allocations.

In practice, the only place where this mapper was used was for saved state, and it re-allocates most nodes anyway because it modifies the locations in them. So, probably not a big deal.

Reviewed By: jbrown215

Differential Revision: D10439250

fbshipit-source-id: f0de9c41623e11fa1682b13db09ca8a57c2155f2