ALoc.none
Summary:
`Loc.none` is used in quite a lot of places (whether it should be used this much is an open question, but not one that I am prepared to address right now). So, to facilitate the conversion to `ALoc`, we need an `ALoc` equivalent.

I can imagine several different representations for `ALoc.none` once we actually change the underlying representation of `ALoc.t`. Basically we need to make sure that `ALoc.of_loc Loc.none` returns `ALoc.none`, and that `ALoc.to_loc ALoc.none` returns `Loc.none`. This is definitely achievable, so I'm comfortable adding this.

Reviewed By: jbrown215

Differential Revision: D10414358

fbshipit-source-id: aaacd441e5a26a35539bebff3fa204fa05a16d11