Functorize `Scope_api` implementation over Loc
Summary:
I'm not yet entirely sure whether I want to convert all users of `Scope_api` over to `ALoc.t`s. For example, find-references uses `Scope_api` and it has no reason to use `ALoc`s. Since `Scope_api` doesn't rely on the structure of locations, it's easy enough to abstract it away and offer two variants for different clients.

As part of this change I removed `scope_api.mli` because adding the signature to the functor accomplishes the same thing and it was painful to avoid excessive code duplication in the `mli` after this change.

Reviewed By: jbrown215

Differential Revision: D10414724

fbshipit-source-id: 9e7211667ed9e70436882f3a5b6f15a5fb3fcc97