[Normalizer] EvalT Spread fix
Summary:
This diff adds `Spread` as an object property in the Ty.t language, and
adds the necessary support in normalization and output.

This allows types like:
```
type T<X: {}> = {... X}
```
to pass normalization. Before, we expected the normalizer to evaluate
spreads and return the resulting object types. This was not possible for
parts that remained abstract. (There is no other way to describe `T` than
the above.) All types of this form would cause the normalizer to return
with an error.

This change allows the normalizer to reconstruct spread objects, whenever
evaluation is not possible, ie. whenever the type contains abstract parts
(`BoundT`s).

Reviewed By: samwgoldman

Differential Revision: D10440889

fbshipit-source-id: 7973b4cdaafc3393cb3093205904ee78cd260726