[type-at-pos] Add implicit instantiation to type-at-pos searcher
Summary:
Adds the implicit instantiation node to the typed-ast searcher for type-at-pos.
Right now it is typed as any, but I will be changing that in a follow-up.

The plan, as discussed previously, is to flow all instantiations of the implicit instantiation to an OpenT and use that OpenT for type-at-pos.

Reviewed By: panagosg7

Differential Revision: D10450598

fbshipit-source-id: 2ee10cd9c612cdf65fe09d9270dd4f24b8fa4093