Profile error collation / formatting in check_once
Summary:
While working on D10232330 I noticed some really confusing performance timings.
The "totalWallTime" metric regressed by a huge amount, but no individual timed
section explained the regression.

Note that the "totalWallTime" regression comes from a timer in the windtunnel
script, which measures the totally time spent executing the test command, which
is `flow check --json --profile`. The rest of the performance data is from our
internal profiling, which captures the durations of nested segments.

As that change introduced new errors, I hypothesized that something to do with
the error printing was responsible, as this section is not already captured by
our internal profiling.

I added CollateErrors in the previous diff, but that did not capture the
regression, although it does partially explain the difference between the
totalWallTime metric and the sum of the internal profiling metrics.

It turns out that printing JSON errors specifically has some performance
issues. Unfortunately, it's a bit awkward to include that computation in a
profiling segment, as the JSON output includes the profiling data itself.

After experimenting with a few alternatives, I settled on this design, which
splits the `print_errors` API into two parts: The `format_errors` function
which does any necessary preprocessing and returns the `print_errors` function
which takes a `Profiling_js.finished` value as its argument and does the rest
of the printing.

I preserved the existing `print_errors` API, which simply calls the
`format_errors` function and then continues to fully apply the returned
closure.

Reviewed By: avikchaudhuri

Differential Revision: D10305070

fbshipit-source-id: 5641703fa57c1a0b3d94ecca0e611b75892c44c0