Propagate any from lower-bound MergedT instead of empty
Summary:
Empty would generally be fine, except that they have a tendency to switch sides
and appear on the right, where they are overly restrictive.

Even the comment above intimates that MergedT is "any with uses" which is
fairly accurate, as MergedT replaces the any that arises from unconstrained
exports.

This will all be moot soon enough when we can finally remove MergedT entirely,
as this stack is working toward.

Reviewed By: avikchaudhuri

Differential Revision: D10419674

fbshipit-source-id: 39f243d091d0ccc6886ce57de1193c3b6f3f6310