Propagate all types to TypeDestructorTriggerT
Summary:
If we prune constraints with TypeDestructorTriggerT as an upper bound, then
whatever was waiting on that trigger will stall.

This comes up in full type resolution, which expects EvalT types to be 0->1. If
the EvalT is processed using TypeDestructorTriggerT, then these stalls can
prevent full type resolution from making progress.

Reviewed By: avikchaudhuri

Differential Revision: D10369368

fbshipit-source-id: 5564823f90cb28af86277d11a30bd55154755e13