Fix type destructor full type resolution
Summary:
An EvalT stores its evaluated type in the context as an optimization. Full type
resolution would fail to wait for an EvalT which was not already evaluated and
contained terms which were not resolved.

It is sufficient to resolve the terms inside the eval type only. If the result
is already evaluated, we will use it. Otherwise, we will have enough resolved
to fully evaluate the term when it appears in a constraint.

Reviewed By: avikchaudhuri

Differential Revision: D10233470

fbshipit-source-id: 070f9882cf234776c250627d9ebb3e69429950c1