[suggest] Do not output <<anonymous class>>
Summary:
Flow suggest should never print `<<anonymous class>>`.

This diff adds an "anonymous" property to `Ty.symbol`, which is initialized when the symbol
is created. For now the above is the only use for this, but perhaps this will be extended to
other cases, where we would like to ban certain symbol names from being output.

This approach should be resilient to codemods that potentially change names of "anonymous"
elements.

The change of `symbol` from a tuple to a record is incidental.

Reviewed By: pakoito

Differential Revision: D10503011

fbshipit-source-id: 384053cd3cee90fb9292bddb1ac5b1d0f345d0de