[hack] Respect --retries hh_client argument while waiting for server to be started due to ongoing rebase
Summary:
When testing D10511311, I've noticed that if `vim` is my merge resolution tool, it can mysteriously hang. This is because they call `hh_client --retries 1` and trust us to return quickly (while also swallowing our error message showing WHY we are not returning).

This was already a problem forever, but I expect with the fixes I make it might start showing up often enough to annoy people.

Reviewed By: pittsw

Differential Revision: D10514653

fbshipit-source-id: 71542bd3669df8bd1cae03e594be14a752d1853e