[Normalizer][Fix] Reset state after normalizing imports
Summary:
File imports are normalized first to determine what the imported symbols of the file are.
These symbols are still considered `Remote` during that time. Before, we used the same
cache after this normalization process which resulted in some imported symbols to still
appear as Remote later during the normal pass.

The normalizer state is reset after normalizing the types of the imports.

Reviewed By: avikchaudhuri

Differential Revision: D10520696

fbshipit-source-id: 622714fc9286dfa2a4c019c9b2e4d3e5dab53717