Be specific about what `Loc_sig` modules are exporting
Summary:
Using `include` here caused some problems with a future diff where a functor took `M: Loc_sig.S` and the return type was some `S with L = M`, where the `S` module type had a module `L : Loc_sig.S`.

The problem arose when attempting to constrain `S` in the `.mli` file to, e.g. `S with Loc_sig.LocS`. Because the module `M` passed to the functor was constrained to have type `Loc_sig.S`, the output type of the functor would only include those functions specified in `Loc_sig.S`. However, because of the `include Loc` in `Loc_sig.LocS`, the `mli` file was inadvertently specifying that *all* functions defined in the `Loc` module should be exposed.

The fix is to just expose the `Loc` functions that we need here. We can't simply hide the unnecessary fields in `LocS` by defining it as `module LocS : Loc_sig.S with type t = Loc.t = struct ... end` because then `LocS.LMap` (and set) would not be compatible with `LocMap` -- all the type system would know is that it is *some* map where the key type is `Loc.t`.

Reviewed By: jbrown215

Differential Revision: D10510646

fbshipit-source-id: d916da31885caf9a009353d05fcdf6e33b1abf44