Add ReadOnly Map/Set/WeakMap/WeakSet
Summary:
Adds new types to core.js to describe readonly Map, Set, WeakMap, and WeakSet. These types partially* follow the pattern of `$ReadOnlyArray` / `Array`, where there is now a "readonly" parent class that has the non-mutating methods, which the regular (mutable) class extends.

The one difference from `$ReadOnlyArray` is that the readonly Map/Set variants are *invariant* on their type parameters. This is an intentional choice for the first iteration of these ReadOnly classes, as it solves an immediate use-case (expressing that a consumer won't/can't modify a value) while preserving the most option to change the API in the future.

There are two high-level approaches that we considered:
* Make the type parameters invariant, as in this diff
* Make the type parameters covariant (all of them - Map keys/values and Set items) but change methods that take type params as input to accept a `mixed` instead (e.g. `Map::get(key: mixed)`). The main downside of this is that Flow would not be able to help detect passing a key that cannot possibly exist in the Map.

Reviewed By: jbrown215

Differential Revision: D10471811

fbshipit-source-id: 2f8f243644ac0d75372d2b6eb80a8ed042852e66