remove incorrect `expression` field from function AST
Summary:
the [`expression` bool on a function node](https://github.com/estree/estree/blob/master/deprecated.md#functions) is a deprecated SpiderMonkey extension to ESTree that refers to ["expression closures"](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Expression_Closures), a non-standard JavaScript 1.8 feature that we don't even parse.

however, esprima seems to use this field anyway. in arrow functions, they set it to true for `() => false` for example, even though that is not an "expression closure". so I kept the field in `Estree_translator` and maintained our current behavior of setting it based on whether the body is a block or expression. but we should probably just get rid of it...

Reviewed By: pakoito

Differential Revision: D10509897

fbshipit-source-id: 61db08434df41d8de4e34725834629c00c0d01ea