[parser] fix bug where variance sigil isn't included in type_param loc
Summary:
One of the unit tests added in D10512084 (`type_alias_param_delete_variance`) fails because the variance sigil in the type param `+A` is not included in the location of type param.

This diff fixes that.

Edit: As of V4 of D10512084, this fix may no longer be necessary for the test to pass. It's still good to fix this though.

Reviewed By: nmote

Differential Revision: D10515549

fbshipit-source-id: 8bc2c8e496629cd8aef12f1b82da062cd5ebbd6d