flowtest package
Summary:
`flowtest` is a CLI tool that you can use to run tests that test your flow types. Any file inside a `__flowtests__` directory or containing the `-flowtest.js` suffix will be checked.

A file will fail the test if it has any errors or warnings. The command will use an already existing flow server to do the checks, if there is one running.

Reviewed By: TheSavior

Differential Revision: D10493244

fbshipit-source-id: 19060fa27428c77ffba1f79ae6d5377f76a68c77