add magic string prefix to upgrade string literals to "module references"
Summary:
This diff introduces a configurable string prefix that will upgrade any string literal in any position into a `$Flow$ModuleRef<$Exports<'TheModule'>>`. This lets currently hard-coded APIs like `requireLazy` be typed in userland, instead of being special-cased in the typechecker.

An example usage is a function like `ifRequired('m#Foo', Foo ==> { /* callback */ });`. This function talks to the underlying `require` internals to actually get the module exports and this new API would let this function be typed like so:

```
declare function ifRequired<TModule, TYes>(
  id: $Flow$ModuleRef<TModule>,
  cb: (module: TModule) => TYes,
);
```

Flow would then understand that the first argument to the callback is in fact the same as requiring `Foo` directly. This reduces boilerplate and `typeof` in some of the magic infrastructural APIs.

Reviewed By: gabelevi

Differential Revision: D10369005

fbshipit-source-id: 75bffb03209003838fd469fa93c122b74b61f790