Add FullyResolved node
Summary:
It's useful at various points to know when a tvar is not only resolved, but
transitively resolved. For example, during full type resolution, we can avoid
walking over the structure of a fully resolved type.

Currently, the only place we know a type is fully resolved is after merging a
component, but in theory we could mark tvars as fully resolved at other points
as well, for example during fully type resolution.

This diff adds the new `FullyResolved t` node variant to the type graph, which
is like `Resolved t` except with the guarantee that `t` is recursively resolved
as well.

Reviewed By: avikchaudhuri

Differential Revision: D10232330

fbshipit-source-id: b2db85010a58139f8b7946e5ed7172426ffa7d69