[Ty] Make properties of Ty.Generic more accurate
Summary:
After cleaning up types abusively called `Generic` earlier we now replace
the `structural` field of the `Generic` constructor with a `gen_kind` field.
This field tracks, *faithfully at this point*, the reason for creating this generic
instance. It's one of class, interface or type-alias. The latter two cases were earlier
classified as having `structural=true`. This enables us to be more precise
in the interaction with constructors like `ClassT`.

Reviewed By: pakoito

Differential Revision: D12838022

fbshipit-source-id: cc0c9cc2a5ae5d3f07914d9da6cef01c0db7e18e