recursive class expression
Summary:
Outlining with a fresh name doesn't currently work when a class expression has a name and the name
appears recursively inside the class expression.

Instead, with this diff, when a class expression is named we will use the name. Using the name is
unsafe when that name is already bound in the toplevel scope, in which case we report an error
during signature verification.

Reviewed By: panagosg7

Differential Revision: D10440397

fbshipit-source-id: 539e91bfd7487b828162aea75c84cf8cf45379ef