remove `flow port` command
Summary:
this command converted a Facebook-specific subset of JSDoc that was used before Flow existed, to Flow annotations.

everything internally has been ported over to Flow annotations, so this code has rotten and costs effort when changing the AST. we recognize there may be some users of this externally, and it may be possible to reimplement in the future, or via a separate tool (e.g. jscodeshift + flow-parser + an existing JSDoc parser).

Reviewed By: gabelevi

Differential Revision: D12882093

fbshipit-source-id: 8364dd40831befc2a21df2d9725046fe048d1fb7