Functorize `Signature_builder_deps` over location
Summary: `File_sig` depends on this, so it is necessary to functorize this in order to functorize `File_sig`. See the next diff for an explanation of why `File_sig` needs to be functorized. The presence of a `Set` means that this must be functorized, and we cannot simply use polymorphic types.

Reviewed By: avikchaudhuri

Differential Revision: D12883968

fbshipit-source-id: 8209c37d6b3f7cf60ef7a5af9a23c7000bbf3148