Functorize `File_sig` over location
Summary:
The parser produces a `(Loc.t, Loc.t) Ast.program` so for consistency it makes sense for it to also produce a `File_sig` with concrete locations. There are also several areas where `File_sig`s with concrete locations are used, so it seems better to avoid unnecessary conversions back to `Loc.t` in those places.

Unfortunately, `File_sig` makes use of `LocSet` so we cannot make it polymorphic: it must be functorized.

Reviewed By: avikchaudhuri

Differential Revision: D12883981

fbshipit-source-id: a8a66bfbedca712f0dbe9adf922d336ee5a0a61d