Update typechecking to use ALoc.t across the board
Summary:
This updates the core typechecking logic to operate solely on structures containing `ALoc.t`s, not `Loc.t`s. This is an important step towards using abstract locations to improve recheck optimizations.

A previous diff addressed the boundary between `ALoc.t`s and `Loc.t`s on the tail end of typechecking (by converting `Error.t` from a version containing `ALoc.t`s to a version containing `Loc.t`s, and then rendering). This diff addresses the boundary on the front end of typechecking, as well as pushing `ALoc.t`s throughout the typechecker.

The strategy here was to convert `AST`s from `(Loc.t, Loc.t) Flow_ast.program` to `(ALoc.t, ALoc.t) Flow_ast.program` right before typechecking, and then chase the compiler errors from there. This also involved converting `File_sig`s over. There were some places where it was less friction to use a `File_sig` with `Loc.t`s, and some where it made more sense to use `ALoc.t`s. As a result, I functorized `File_sig` (see D12883981) and created a conversion function to use where needed.

Note that all of the conversions from `Loc.t` to `ALoc.t` in this diff are trivial and will remain so. Even once we have actual abstract locations, the underlying representation of many locations will simply be a concrete `Loc.t`. All of the conversions introduced in this diff will be conversions from `Loc.t` to an `ALoc.t` with a concrete representation. The step that actually creates abstract locations with keys and such will be self-contained, and will operate on the output of the signature builder step.

I was concerned that the extra work done when we convert ASTs and `File_sig`s would negatively impact perf. When we do this, we have to traverse the entire structure and re-allocate all (or most) nodes. However, it appears that this perf cost is negligible.

This diff accomplishes a lot. However, there are still some followup items:
* I made a mistake by having the type inference hooks continue to operate on `Loc.t`. This design requires a lot of ad-hoc conversions to `Loc.t` in places that should not be doing such conversions. By the time I realized this was a mistake I was far enough down the path that I chose to just finish it and deal with it later. I'll fix this shortly in a followup diff.
* The `ESignatureVerification` error should most likely use a structure containing `ALoc.t`s. By the time I realized this, the diff was already building and I didn't want to make it any larger than it already is.
* I'm not sure whether `error_suppressions.ml` should use `Loc.t`s or not. Right now it still does. I'll have to think about this a bit.
* `Typed_ast_utils` is a bit of a mess now with an ad-hoc mix of `Loc.t`s and `ALoc.t`s. I need to spend a bit of time cleaning it up.

Reviewed By: avikchaudhuri

Differential Revision: D12769364

fbshipit-source-id: 90f813afe60505ebbd07cf866f01993cfeed2eab