Fix various mappers not walking the expression of a computed object key
Summary:
This fixes various mappers that were not explicitly walking the `key` of an object property when it was "computed". (Something like `{[x]: y}`. This meant that arbitrary expressions were not processed by these mappers, which led to problems.

```
let foo = "name";
let obj = {[foo]: 1};
```

When renaming the renamer would not walk the computed property and so it would end up like this:

```
let a = "name";
let b = {[foo]: 1};
```

And crash.

This also fixes the es3 guard to throw over computed properties.

Reviewed By: mroch

Differential Revision: D12901598

fbshipit-source-id: 49803bb57677d1463ba1ef3d44d7b5d4bd254eb2