[AST] remove redundant Expression.Member.computed field
Summary:
a member expression is computed if its property is a `PropertyExpression` vs a `PropertyIdentifier`. a separate `computed` field is redundant and allows bogus ASTs.

this flag is useful in estree to differentiate between `foo.bar` and `foo[bar]`, since the property is an identifier either way, but the ocaml ast uses a variant instead.

Reviewed By: nmote

Differential Revision: D12924869

fbshipit-source-id: 0b72b99b59c703f3d9a607c0c650d0568216fd39