allow configuring path to node during runtests.sh
Summary:
the custom resolver tests are relying on node from the path, but it can be helpful to override the path.

this diff introduces a `FLOW_NODE_BINARY` environment variable which you can point at a custom node:

```
FLOW_NODE_BINARY=path/to/node ./runtests.sh path/to/flow
```

This is accomplished via a wrapper script, `scripts/tests_bin/node`, which calls `$FLOW_NODE_BINARY`. `runtests.sh` sets the var to `NODE_BINARY` if it's set, or `which node` if not. `scripts/tests_bin` is added to `PATH` when running the tests, so `node` in tests picks up the wrapper.

Reviewed By: gabelevi

Differential Revision: D12964367

fbshipit-source-id: ef98fb1f89c6a32f04ef36ebb315a604f01c5e9e