Store errors from signature builder with ALocs rather than Locs
Summary:
This probably doesn't matter a whole lot either way since these come fairly directly from the AST and cannot change without the file in question changing. However, it's more consistent with the rest of the errors to use `ALoc.t` here.

This is a fairly mechanical change that just converts to `ALoc` a bit earlier than we did previously.

Reviewed By: jbrown215

Differential Revision: D12912779

fbshipit-source-id: fb43365f3b10e17930d9945e54c16cfa5766387e