[Normalizer] Handle import errors more finely
Summary:
The normalizer precomputes all imported names, in order to later
determine what symbols have properly been imported.

Before, we required *all* imports of a file to be normalized without
errors to create the imports map. This change lifts this restriction,
by instead collecting a map of the imports that have normalized correctly
and just ignoring the rest.

Reviewed By: samwgoldman, pakoito

Differential Revision: D12913583

fbshipit-source-id: 77f3d98b3cbb90649904dfcee41eab4b452ae88f