[komodo] Only normalize once per file
Summary:
Before, we would normalize all locations of a file F for each query originating from F.
This led to several redundant normalization passes. This diff removes this redundancy
by calling normalization once per file.

In addition it parameterizes komodo-mapper with the options to the normalizer, instead of
hardcoding the same options for every codemod.

Reviewed By: pakoito

Differential Revision: D12953695

fbshipit-source-id: cc8bfcdcd7f594dd5219ba970ae59e403a945f4a