[react] reserve React$AbstractComponent
Summary:
Adds React$AbstractComponent as a built-in that expects 3 type arguments. I'll be building out the functionality of this type in stacked diffs.

OSS Users: I will follow up with a blog post when we get closer to having this type actually be usable. We expect it to improve the precision and speed with which flow can typecheck React components.

Reviewed By: samwgoldman

Differential Revision: D12925858

fbshipit-source-id: e242dbf74ed2b90eca3fb7f66641bee8b919e530