Remove weird mixed case from LookupT
Summary:
A while back I added a TODO comment asking why we would ever hit MixedT in a
LookupT failure. I'm finally coming back to figure out why and clean up this
code.

The lower bound of a LookupT is supposed to be something on the prototype
chain of an object or an instance, where a GetPropT or SetPropT has already
failed. Unlike GetPropT/SetPropT, LookupT constraints are not syntax-directed.
The significance of thise is that NullT as a lower bound of LookupT always
means a null prototype, rather than a null value.

This also means that a MixedT lower bound is generally unexpected, as a
prototype chain should only include object-like things and null.

It turns out that this case is only significant in the TestPropT case, which
forwards all non-null lower bounds to LookupT. This means that the lower bound
of a LookupT can be mixed. For example, in the program `if (o && o.foo) {}`
given that `o` is mixed.

It is simpler to handle this case in TestPropT directly, which is what this
diff does. I have also added a comment explaining why this logic is safe.

Reviewed By: panagosg7

Differential Revision: D13036543

fbshipit-source-id: 9e9f2db5137508958fc0ba262f2bf3221a39958b