[type-at-pos] Update eval tests
Summary:
This change only prepares the type-at-pos tests to avoid noise in the subsequent diff.

The code appended in the end of tests/type-at-pos_types/spread.js ensures that the relevant
EvalTs are evaluated by Flow's main checking routine. In the next diff, we'll be changing the
normalizer so that it does not trigger any further constraint generation. So here we're only
emulating this process, to keep the test as faithful to it's original purpose as possible.

The changes in the beginning of spread.js remove flow errors.

Reviewed By: avikchaudhuri

Differential Revision: D13079194

fbshipit-source-id: 728375fba5fc26c7f18a4e143611451c48e77f04