fix file_sig require processing for generated signature ASTs
Summary:
Requires and, more generally, other constructs are "normalized" in generated signature ASTs, which
means that they might be split into elementary parts, which might in turn cause some of the
corresponding AST nodes to appear multiple times in the AST.

Processing requires in `File_sig` involves a complicated visitor that visits some nodes multiple
times, and relies on some caching involving the locations of those nodes. Unfortunately when the
same nodes can appear multiple times, the caching mechanism breaks.

This diff exploits knowledge of the different paths through the visitor that cause multiple visits
to the same node to fix the caching mechanism.

Reviewed By: panagosg7

Differential Revision: D12972962

fbshipit-source-id: 982f066a15d63f5c842f7de33a431b248c515d34