replace AnyObjT with AnyT with Object source
Summary: kill AnyObjT, replace with AnyT AnyObject. Ideally we would not depend on the specific source of an any to determine behavior but this will come in another diff. This has two main semantic effects: `Object` types are now interpreted as `any`, so we output them as such for things like type-at-pos. We also now allow some extra behavior like `let x : Object = 4` which was previously disallowed.

Reviewed By: avikchaudhuri

Differential Revision: D12874952

fbshipit-source-id: bd383ab67cf3386b5e42413cc1d7d1eaa8b028e5