replace AnyFunT with AnyT with Function source
Summary: kill AnyFunT, replace with AnyT AnyFunction. Ideally we would not depend on the specific source of an any to determine behavior but this will come in another diff. This has two main semantic effects: Function types are now interpreted as any, so we output them as such for things like type-at-pos. We also now allow some extra behavior like let x : Function = 4 which was previously disallowed.

Reviewed By: avikchaudhuri

Differential Revision: D13003036

fbshipit-source-id: b8ab80a5e44f82d4063844799f1c17415246c9a5