Fix bug in polymorphic AST mapper
Summary:
The mapper incorrectly set the `shorthand` field of object property patterns to `false`. This caused a change in test output for the `rename` test when combined with the next diff (D13095655).

We should probably have some principled testing for the polymorphic AST mapper. Maybe run it over a large codebase and ensure that for each file, pretty printing the before and after ASTs gives the same result.

Reviewed By: panagosg7

Differential Revision: D13105099

fbshipit-source-id: 46e6d26b1fe0ac5e9dddf84040db190e27925590