[hack] Watchman.with_instance
Summary:
Users of the `Watchman` API usually have a `watchman_instance`, which is either `Watchman_alive` or `Watchman_dead`. There's already an internal API `call_on_instance`, but it makes assumptions about the return type.

So `with_instance` is a more general API. I use it for closing connections in {D13060170} and to get the mergebase in {D13060167}

Reviewed By: jbrown215

Differential Revision: D13060168

fbshipit-source-id: 3dadff56394497d85cb8eb087ca6412ec70419f2