[hack] Watchman.get_changes_since_mergebase
Summary:
This adds an API to `Watchman` and `Watchman_lwt` which uses a one-off scm-aware watchman query to figure out which files have changed since the mergebase with master.

I'm building a lazy mode into Flow which, upon starting up, will recheck the files which have changed since the mergebase.

It's important to note, this will NOT use the existing connection for this one-off query. The existing connection might be used for a subscription, which means watchman might push a message to us and we may mis-interpret that message as a response.

Reviewed By: jbrown215

Differential Revision: D13060166

fbshipit-source-id: 32bf67d18f1d691a3e19e74550615b039f881c12