--lazy-mode watchman
Summary:
This introduces a new lazy mode named `"watchman"`. It's basically the same thing as `--lazy-mode fs`. That is, whenever a file changes we focus on that file and recheck it and its dependents & dependencies.

The big difference is that `--lazy-mode watchman` will query watchman at start-up time for the files which have changed since the mergebase with master. This should give us a pretty good estimate of which files the user cares about.

So whereas restarting a server in `--lazy-mode fs` will cause all files to become unfocused, restarting a server in `--lazy-mode watchman` should allow us to re-focus on all the files the user cares about.

Note - when figuring out what files changed since the mergebase, we filter out files which would cause Flow to restart when rechecked (e.g `.flowconfig`, libdefs, package.json, etc). These files haven't changed since the server initialized, so there's no need to restart.

Reviewed By: jbrown215

Differential Revision: D13060162

fbshipit-source-id: a43e6f0daff2b25abb22efe95e2bf0b3552393df