Thread metadata from Watchman into recheck
Summary:
Watchman can tell us cool things in addition to which files changed. This info can help us in the future decide whether to do a recheck or to restart the server.

This diff threads the metadata that Watchman collects in the monitor process through to the server's rechecking code. Currently this metadata includes whether the mergebase with master has changed (though as of this diff it's always false) and how far we've rebased in total.

As of this diff the server doesn't currently do anything with this info other than log it.

Reviewed By: jbrown215

Differential Revision: D13060169

fbshipit-source-id: 6d539875df5a0d0c67c4c945e843f35c2cd974a7